requirements.txt
yookassa~=2.3.5
python-dotenv~=0.21.1
aiogram~=3.2.0
asyncpg~=0.27.0
APScheduler~=3.10.1
tzlocal~=4.2
regex==2022.10.31
